refactor#359: onSuccess, onError 라이프사이클 활용
